Generate Expr(:incomplete) for errors which hit EOF (JuliaLang/JuliaSyntax.jl#102)
This allows REPL completion to work correctly. It works by pattern matching the parse tree, rather than hard coding incomplete expression detection into the parser itself. There's still more changes from JuliaLang/JuliaSyntax.jl#88 which would help make this nicer but for now it works.

The tests here are somewhat derived from Base, but had to be reviewed and tweaked because they turn out to not really be consistent. For example, "begin;" and "begin" are both the prefix of a block construct, one of them shouldn't come out as `:other`.