Update interfaces.md (#46833)
AbstractArrays have never been required to be mutable, and we have many subtypes of `AbstractArray` in `Base` and `LinearAlgebra` that do not implement `setindex!` since they are immutable, or only implement it for a subset of their indices (e.g. `CartesianIndices`, the various range types, `Symmetric`, `Hermitian`, `Diagonal`, ). There is also a lot of hope and plans for more immutable, dynamic array support from Base, and of course a rich set of immutable arrays out there in the package ecosystem which do not have `setindex!`