Inlining: Remove outdated code path for GlobalRef movement
We used to not allow GlobalRef in PhiNode at all (because they
could have side effects). However, we then change the IR to make
side-effecting GlobalRefs illegal in statement position in general,
so now PhiNodes values are just regular value position, so there's
no reason any more to try to move GlobalRefs out to statement position
in inlining. Moreover, doing so introduces a bunch of unnecesary
GlobalRefs that weren't being moved back. We could fix that separately
by setting appropriate flags, but it's simpler to just get rid
of this special case entirely.