irinterp: Add a hook for control-dependent phi evaluation (#46883)
External AbstractInterpreters may want to modify the lattice elements
of a phi to track additional information (e.g. by inspecting the branch
condition of the idom). Currently, we don't pass enough information
down to make this happen, so add a hook function that also takes (ir, idx)
and a lazy domtree. Admittedly this is a bit ad-hoc, and would probably
be better with some sort of general analysis manager, but this'll do for
now.