precompile: serialize the full edges graph
Previously, we would flatten the edges graph during serialization, to
simplify the deserialization codes, but that now was adding complexity
and confusion and uncertainty to the code paths. Clean that all up, so
that we do not do that. This also avoids needing to interact with
backedges at all.

This uses the same algorithm now as #46749 for cycle convergence.