Reorder FileEvent field names in the documentation (#46905)
When printing a FileEvent the user sees a text in the format `"path" => FileWatching.FileEvent(bool, bool, bool)` and by reading the documentation the user expect the three booleans to correspond to `changed`, `renamed`, `timedout` while instead the correct order is `renamed`, `changed`, `timedout`. This is just a correction to that.