gf: detect more ambiguities
As long as we find at least one other method that conflicts fully, we
know they will exist together for all subtypes, and always be ambiguous
for all subtypes too. This is already the algorithm used by method
insertion to decide if the new method is ambiguous with a "missing"
method.

Fix #46601