precompile: serialize the full edges graph (#46920)
Previously, we would flatten the edges graph during serialization, to
simplify the deserialization codes, but that now was adding complexity
and confusion and uncertainty to the code paths. Clean that all up, so
that we do not do that. Validation is performed while they are represented
as forward edges, so avoids needing to interact with backedges at all.

This uses the same algorithm now as #46749 for cycle convergence.

(cherry picked from commit fbd5a72b0a13223c0f67b9d2baa4428e7db199a5)