Fix oracle check failure in adce_pass
This is probably a real issue, though I did not see it cause an actual issue,
only an oracle check failure if the verification is turned on. The issue
was that we were failing to count the removal of phi node edges during
the adce pass, so we were left with excessive counts at completion.