Special purpose inliner for Core.ifelse
This function isn't used as much anymore now that Base.ifelse is
an actual generic function, but it's still used in a few places
across the ecosystem, and there currently isn't anything that
would fold it for constant conditions, so add a special case
inliner for it. This probably doesn't have a huge impact, but
I happened to run into a case where it was causing annoying
suboptimialities and it's a quick fix.