Make @specialize actually work if the module was @nospecialize'd (#47106)
In compiler/tfuncs.jl, we're assuming that after a `@nospecialize`
at top level, a subsequent `@specialize` on the argument will
override the toplevel `@nospecialize` (just like an argument-level
`@nospecialize` overrides a toplevel `@specialize`). However,
it turns out we never implemented that support, so the annotation
was silently ignored. Fix that and add a test.