Do not use Stateful for cmp of AbstractString
PR #45924 fixed length of Stateful, but this change requires Stateful to call
length on its wrapped iterator on instantiation.
The current implementation of cmp(::AbstractString, ::AbstractString) wraps the
strings in Stateful to efficiently check if one string is longer than the other
without needing an O(N) call to length.
However, with #45924 this length call is done anyway, which led to a performance
regression.

In this PR, the cmp method is changed so it no longer relies on Stateful to do
the same thing.

Fix #46719