Fix various bugs in maybe-undef handling
We weren't accounting for the possiblity that a slot could be undef
in our effect handling and the recently introduced shortcut in the
type lifting pass wasn't quite correct. Ironically, the test case
I have here compiles correctly on master, because the compiler
miscompiles itself (due to this issue) to always make the `@isdefined(last_val)`
check `false` in this test case. We can still look at the IR though,
which this does.

Fixes #47127