Also update phi counts in adce pass
The whole point of this pass is to compute and compare the
counts of all SSA value uses vs those of only-phi uses to
find SSA values that have no real uses. In #47080, I updated
the code to properly account for removal of phi edges in
the SSA count, but neglected to do the same in the phi-only
count, leading to #47180. Fix that.

Fixes #47180