remove kwfuncs from Julia (#47157)
* remove kwfuncs from Julia

These kwsorter methods only exist for dispatch, which means they do not
need to be unique. We can optionally have the primary MethodTable
contain an extra kwtable::MethodTable field, if this turns out to be
slow, since this already introduces the concept of
`jl_kwmethod_table_for` (for better reflection and max_args).

* remove jl_f_kwinvoke builtin, reimplement in Julia

This instantly grants total inference and inlining support, where
previously this was a completely opaque call!