Fix insert_node! for insertion at pending nodes
We didn't handle this at all, now we at least handle insertions
with `attach_after` an existing node. Attaching before is
currently disallowed. We could have it attach before the node to which it is
being attached, but it is not clear that this is what we want. We can
revisit that if somebody has a usecase where they want that.