Add option to only install essential signal handlers
When embedding Julia, it is useful to opt out of some of the signal
handlers (especially SIGINT). But opting out of SIGSEGV makes it very
difficult to use multiple Julia threads, since the GC safepoint
implementation relies on segv_handler().

This patch provides a third option to embedders, which installs the
essential signal handlers but skips the optional ones. It might make
sense for this to be the default when embedding, but I've kept the other
two options for backwards compatibility.