Add utility `IRCode()` constructor (#47218)
Just to make it easy to get one. Right now, people tend to do `code_ircode` on
something random and then strip it which is just silly. Should mostly be used
in debug and utility code. Proper compile pipelines should still get it from
the frontend or the cache, but it seems like a useful addition.