cfg_simplify: Fix more bugs (#47240)
cfg_simplify still had issues with unreachable BBs, as well as BBs with
try/catch in them. Additionally, it sometimes leaves unreachable BBs
in the IR, which the verifier was upset about if there was a PhiNode
that referenced it. I made that legal for now. The alternative is to
make all unreachable BBs illegal, which I think would be reasonable,
but is somewhat extreme for the time being. Let's see how this
fares first.