Make sure to preserve info in union-split (#47301)
Generally we try pretty hard in inlining to make sure to preserve
the :info of a statement when we turn it into :invoke. We don't
use this information in the compiler itself, but it is used
by external tools (including Cthulhu). However, we were dropping
this info during union splitting for no particular reason. Fix
that.