Fix yet more cfg_simplify bugs (#47286)
I continue to throw cfg_simplify at big external IR examples and
unfortunately, I continue finding bugs. Hopefully this is the last
of it. The particular bug here had to do with scheduling fallthroughs
properly across dropped basic blocks. The test gives an example,
but it's pretty hard to hit this case.