Fix const struct field errors + cleanup `global const` AST (JuliaLang/JuliaSyntax.jl#130)
Here I've replicated the fix from JuliaLang/julia#45024 so that
`const x` (ie, without an assignment) is only valid within a `struct`
and is otherwise an error.

Also avoid lowering the syntax `global const` into `const global` within
the parser; do this in `Expr` conversion instead. This more closely
reflects the structure of the source, allowing trivia attachment to be
more natural.