Improve error message for `which(fn, types)` and `MethodError`
The `MethodError` printing has detailed text on what failed about this
dispatch, which can be useful to someone trying to understand why
`which` failed, given that `which` is simply a model of the dispatch
done by `invoke`. Reuse that printing, but add a short custom header to
it and convert to an ErrorException, per triage discussion that this
would be clearest. Also add more text to `MethodError` to guide new
users in interpreting the exact meaning of some of the cryptic text.

Fixes #47322