Change confusing (likely wrong) docs
At LLVM: "No Infs - Allow optimizations to assume the arguments and result are not +/-Inf. If an argument is +/-Inf, or the result would be +/-Inf, it produces a poison value instead."

I believe a Float64 can't return a posion value (e.g. it wouldn't mean NaN), so doesn't apply for Julia.

[skip ci]