Fixes for macro paths and call chain parsing (JuliaLang/JuliaSyntax.jl#144)
Correctly parse dots and other same-precedence chaining after macro
calls:

    @A().x
    @A{a}.x
    @A[a].x

Also allow general forms in macro paths, such as calls:

    f().@x

This is allowed by the reference parser but seemed weird so it was
disallowed here. But it turns out this doesn't prevent computed macro
paths as the `A` in `A.@x` is looked up dynamically - so just simplify
things and allow this.