use `Base.return_types` instead of `Core.Compiler.return_type`
in the "inference" test suite, since the behavior of the latter is
pretty much undefined and usually we want to use the former for testing
the precision of type inference.

Also fixes up the implementation of `Base.return_types` so that it
doesn't miss to call `widenconst` when inferring a builtin call.