add `isunix` for BinaryPlatforms `AbstractPlatform`
This provides symmetry with the methods defined in `Sys`,
and is convenient for use in e.g. BinaryBuilder platform filters.
Note that there are other `is.*bsd` variants we omit since they
are unsupported by BinaryPlatforms. The capitilization
conventions are also different so we do not call the Sys methods
directly.