faster `BigFloat(::Float64)`
This technically has different behavior than the current behavior but I'm pretty sure that no one was using `BigFloat` with precision<53 because the current behavior is to give you all 53 bits for precision>=40 and only the leading 8 bits for precision<40. This is roughly 2x faster for precision<10000. It's a little slower for precisions >10^6 (about 20%) but IMO this doesn't really matter since for high precision things like multiplication will be way slower than construction.