Enable coredumps on Windows
This commit provides two pieces; first, it enables building libuv with
`SIGQUIT` support on Windows (useful for [0]), and second it enables
coredumping when exceptions are hit, after our exception handler has
finished printing out a backtrace.  This is useful for using WER to
create dump files when we segfault, etc...

This commit requires libuv PRs [1] and [2] to properly function.

[0] https://github.com/JuliaLang/julia/pull/45864
[1] https://github.com/JuliaLang/libuv/pull/30
[2] https://github.com/JuliaLang/libuv/pull/31