Validate literals at parse time (JuliaLang/JuliaSyntax.jl#149)
Here we unconditionally validate all literals after `parse!()` - ie
processing String and Char escape sequences and detecting numeric
overflow.

Doing this up front in `parse!()` means it'll be redone later during
conversion to SyntaxNode. The time cost of this in parsing to Expr seems
to be about 20% in the worst case where the code consists of a large
array of Float64 literals. Which isn't great but is probably acceptable
for now.