Add an inference option to assume no new bindings being added (#47674)
This is currently intended for use with external AbstractInterpreters
that want to be able to constant fold `:isdefined` away. We always
support the `true` case, since we don't support removing bindings,
but the `false` case is currently not foldable. In the future,
we might partition bindings by world age, in which case we would
likely get this for free, but for now, just add this as a hook for
external AbstractInterpreters to use.