Change `@test ex broken=true` to error if `ex` evaluates to non-boolean
Just like `@test ex` errors if `ex` does not evaluable to a boolean, the
`broken=true` variant should as well.  As it stands, right now the
following does not result in an error:

```
@test 1 broken=true
```

While the following does result in an error:
```
@test 1
```

The intention of `broken=true` is to allow for tests to be added such
that they pass although something is broken, and then fail once the
missing functionality has been added, so that the developer knows to go
and adjust the tests for the newly-added functionality.  It is often
used in situations such as:

```
@test partially_implemented_test() broken=Sys.iswindows()
```

Sometimes, the function under test throws an error in certain
situations, and this test can even be used to track a function erroring
out, and notifying the developer when it no longer throws:

```
@test this_throws() broken=true
```

However, this occasionally leads to improper usage such as:
```
@test this_usually_returns_an_object_but_sometimes_throws() broken=true
```

This test, when it throws, passes (because of `broken=true`) but then
when it returns a non-boolean, it _also_ passes.  This is Very Bad (TM).
This PR fixes the state of affairs by making a non-boolean return an
error here.