Fix missing GC root in Symbol construction
The `Symbol` constructor in boot.jl was not using the unsafe_convert mechanism,
becuase it is unavailable at this point in bootstrap. However, it was also not
GC-rooting the string some other way, resulting in potential memory corruption.
Fix that by manually inlining the :foreigncall and setting up the root
appropriately.