Remove wrong `@test_broken @inferred` test
This test does not fail on any Julia version I've tested going all the
way back to v0.7 when it was first introduced.  My assumption is that it
broke on an RC version, and we didn't noticed when it was fixed because
`@test_broken` ignored tests that returned non-boolean values.