isassigned: remove isempty check
This check was added in the previous commit to preserve existing
behavior when calling isassigned with an empty splat for the indexes.
But it turns out different version of isassigned treat this case
differently (e.g. AbstractArray vs Array). So the check is being
removed to make isassigned behavior consistent and more in line
with checkbounds and expectations of the [] operator.