Add option to `jl_print_task_backtraces(bool)` to control whether or not to print DONE tasks (#47933)
Add option to `jl_print_task_backtraces(bool)` to control whether or not to print DONE tasks

This lets you print all live tasks, with less noise in the logs, in case
it's been a while since the last GC.

In some cases, we are seeing many thousands of DONE Tasks, which greatly
increase the noise in the logs, and can overwhelm the DataDog logging.

The API now takes a bool, which can allow us to control this setting, letting us print
the DONE tasks if the user wants to, but also hide them if desired.

(cherry picked from commit 83245120b7cb79b10a6a21c18fae48c7e7ae2c88)