ir: Respect GlobalRef lattice elements
Currently IncrementalCompact recomputes the type of globals on every iteration.
There is not much reason to do this - the type of a global cannot change.
In addition, external abstract interpreters may want to inject custom, more
precise lattice elements for globals, which should be respected.
Overall, this should be both faster and better for external absint,
though of course GlobalRefs now need to be inserted into the IR with the
correct type. If there's any callsites that don't do that, those would
have to be updated.